remove a virtualenv by uuid
consider --rm an expert option, requiring the user to provide to uuid of
the virtualenv to remove. This makes this use case less error prone as
there is no risk in detecting an incorrect virtualenv to remove. Also,
there is no need to create the virtualenv in the first place if it
didn't exist yet.